Pundit with image_results finished but commeted out

+++ app/controllers/comments_controller.rb
@@ -3,6 +3,7 @@ class CommentsController < ApplicationController
   def new
     @post = Post.find(params[:post_id])
     @comment = Comment.new
+    authorize @comment
   end
 
   def create
@@ -10,6 +11,7 @@ class CommentsController < ApplicationController
     @comment = Comment.new(comment_params)
     @comment.user = current_user
     @comment.post = @post
+    authorize @comment
     if @comment.save
       redirect_to post_path(@post)
     else
@@ -21,11 +23,13 @@ class CommentsController < ApplicationController
 
   def edit
     @comment = Comment.find(params[:id])
+    authorize @comment
   end
 
   def update
     @comment = Comment.find(params[:id])
     @post = comment.post
+    authorize @comment
     if @comment.update(comment_params)
       redirect_to post_path(@post)
     else
@@ -36,6 +40,7 @@ class CommentsController < ApplicationController
   def destroy
     @comment = Comment.find(params[:id])
     @post = @comment.post
+    authorize @comment
     if @comment.destroy
       redirect_to post_path(@post)
     end
